requirements.txt
pdfplumber
psycopg2-binary
fastapi
python-multipart